Add an enum translator and start exposing Stim v.1.13 Python API
We add a new enum translator that works in a dynamic setting - as far as we can
tell, CLBIND_TRANSLATE_SYMBOL_TO_ENUM isn't possible to use in this dynamic
setting.

We start to expose the API more systematically by taking the API reference
https://github.com/quantumlib/Stim/wiki/Stim-v1.13-API-Reference, and for each
function in the reference, declare one or more C++ functions that are to be
exposed by clbind.

This way, it is less likely that some functionality will be missed.