Start to expose the API methods in a more systematic way.
There are 392 API methods to expose. Each one can be exposed by uncommenting its
corresponding 'def' statement in stim_clbind.cc, and uncommenting its definition
in stim_clasp.cc and filling in the body.

It is anticipated that most of them will be rather short and can be copied
easily from the pybind.cc files that make up the Stim Python API.